Fold quantile helpers into test_quantile header
